requirements.txt
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
numpy==1.21.0
pandas==1.1.5
PyMuPDF==1.17.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.26.0
tqdm==4.46.1
joblib==0.16.0
sklearn_crfsuite==0.3.6
lxml==4.6.5
pytest==6.0.2
pytest-cov==2.10.1
pytest-mock==3.3.1
coveralls==3.3.1
requests-mock==1.9.3
pylint==2.11.1
scikit-learn==0.24.2
nltk==3.6.6